requirements.txt
streamlit
yfinance
numpy
pandas
plotly
scipy
scikit-learn
fpdf
openai
tensorflow
transformers
torch
newsapi-python
python-dotenv
stable-baselines3
gymnasium
kaleido


